Get latest iOS SDK
