+++ stripe-react-native.podspec
@@ -2,7 +2,7 @@ require 'json'
 
 package = JSON.parse(File.read(File.join(__dir__, 'package.json')))
 # Keep stripe_version in sync with https://github.com/stripe/stripe-identity-react-native/blob/main/stripe-identity-react-native.podspec
-stripe_version = '~> 25.11.0'
+stripe_version = '~> 25.12.0'
 
 fabric_enabled = ENV['RCT_NEW_ARCH_ENABLED'] == '1'
